Move common commands handler from common.py to command.py for Telegram backend
